Add configData
Add necessary dependencies to the pom.
Change the application.yml files to bootstrap.yml adding the configuration so that the microservices can be identified with config-server. The internal name must match the name of the configData file.
The contents of the application.yml files of each microservice were redirected to configData with their respective names.